requirements.txt
discord.py>=2.4
aiohttp>=3.8.0
Pillow>=10.0.0
PyGithub>=2.1.0
requests>=2.32.0
python-dotenv>=1.0.0
